pyproject.toml
[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"

[project]
name = "homebrewtools"
version = "0.1"
description = "My homebrewtools"
authors = [{name="Morten Brekkevold", email="morten@brekkevold.no"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "Pillow",
]

[project.urls]
"Homepage" = "https://github.com/lunkwill42/homebrewtools/"

[project.scripts]
labeltile = "homebrewtools.labeltile:main"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]

[tool.black]
line-length = 120
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.nox
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''
